mass conversion of libicv to using size_t for image dimensions and other size-based values.  warrants more testing, but not seeing any regressions first pass through.
svn:revision:64599
svn:branch:trunk
svn:account:brlcad

+++ src/libicv/color_space.c
@@ -38,8 +38,8 @@ icv_gray2rgb(icv_image_t *img)
 {
     double *out_data, *op;
     double *in_data;
-    long int size;
-    long int i = 0;
+    size_t size;
+    size_t i = 0;
 
     ICV_IMAGE_VAL_INT(img);
